feat: CoreData save, delete 로직 변경 #19
- predicate 변경
- 버튼 바인딩 로직 구현

+++ Left7/Domain/Usecase/YogiDetailUsecase.swift
@@ -14,11 +14,15 @@ final class YogiDetailUsecase {
         self.favoriteProductRepository = favoriteProductRepository
     }
     
-    func saveFavoriteProduct(_ product: Product) {
-        favoriteProductRepository.saveFavoriteProduct(product)
-    }
-    
-    func deleteFavoriteProduct(_ product: Product) {
-        favoriteProductRepository.deleteFavoriteProduct(product)
+    func updateFavoriteProduct(_ product: Product?) {
+        guard let product = product else {
+            return
+        }
+        
+        if product.isFavorite == true {
+            favoriteProductRepository.saveFavoriteProduct(product)
+        } else {
+            favoriteProductRepository.deleteFavoriteProduct(product)
+        }
     }
 }
